Refactored config objects with go-flags
This makes command line flags more easily defined as the go-flags
library and can be more easily parsed, which simplifies the code.

Further, if we want to eventually construct mybench runs via alternative
ways to specify configurations (like via a file, or via a RPC from some
sort of distributed mybench), using this library will give us that
flexibility.

The downside is that all the benchmarks needs to change, and how we run
mybench needs to change slightly (from a single dash to two dashes).