Removed `BenchmarkApp[ConfigT]` and introduced `BenchmarkInterface`
Previously, to define a custom benchmark, you need to create a
`BenchmarkApp[ConfigT]` object via the `mybench.NewBenchmarkApp`
function. You would need to pass the name, the config, and two functions
(`runLoader`, `setupBenchmark`). This is a somewhat clunky interface,
and it relied on a generic (`ConfigT`) for a custom configuration type
unnecessarily. This made the initialization code a bit confusing to read
and write.

With this change, we introduced an interface called `BenchmarkInterface`
and completely removed `BenchmarkApp[ConfigT]`. To implement a
`BenchmarkInterface`, you need to define the name, config, and two
functions (`RunLoader` and `Workloads`). This preserves all behaviors
from the original `BenchmarkApp`, but removed the 2 - 3 structs with
generics that we previously had to go through to get the same
functionality. The old and confusing `setupBenchmark` function has now
been replaced by a `Workloads` function that simply returns a list of
workloads the benchmark is supposed to mix together.

Once the `BenchmarkInterface` is implemented, one can simply run the
benchmark via `mybench.Run(benchmarkInterface)`. In the future, it is
possible that we can dynamically start different benchmarks via this API
call (such as for distributed mybench), as it simply requires a struct
that implements `BenchmarkInterface`.

As a note, `BenchmarkInterface` is conceptually similar to the
`WorkloadInterface`. While `WorkloadInterface` is user-defined to
contain the code for the workload, the `BenchmarkInterface` is
user-defined to contain multiple workloads.