Removed selfbench
This test benchmark was originally intended to test the inner workings
of the mybench parallism. However, the workload here is CPU bound. A
system cannot scale to more CPU-bound goroutines than the number of CPU
cores it has. Most mybench workloads should be IO bound, which means it
can scale to more goroutines than CPU cores. This makes the benchmark
not realistic and effectively useless. It is thus removed.

The code here could theoretically be useful for testing the mybench
looper in some sort of unit/integration testing. However, there's no
reason to keep this as is and it is always possible to resurrect this
code if needed.