Renamed the JavaScript vendor folder to thirdparty
It seems like when you `go get github.com/Shopify/mybench`, `go get` is
likely removing every `vendor` folders recursively in the repo. This is
causing the VegaLite JavaScript to not be included. This means the web
UI is broken for any custom benchmarks in another repo that imports
mybench.